requirements.txt
asgiref==3.9.1
asttokens==3.0.0
decorator==5.2.1
Django==5.2.5
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.0
executing==2.2.1
ipython==8.37.0
jedi==0.19.2
matplotlib-inline==0.1.7
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
wcwidth==0.2.13
